+++ src/mailman/model/bounce.py
@@ -20,8 +20,9 @@
 import logging
 import datetime
 
-from lazr.config import as_boolean
-from mailman.app.bounces import send_probe
+from email.utils import make_msgid
+from lazr.config import as_boolean, as_timedelta
+from mailman.app.bounces import PENDABLE_LIFETIME, _ProbePendable, send_probe
 from mailman.app.membership import delete_member
 from mailman.app.notifications import (
     send_admin_disable_notice, send_admin_increment_notice,
@@ -34,6 +35,8 @@ from mailman.interfaces.bounce import (
     BounceContext, IBounceEvent, IBounceProcessor, InvalidBounceEvent)
 from mailman.interfaces.listmanager import IListManager
 from mailman.interfaces.member import DeliveryStatus, IMembershipManager
+from mailman.interfaces.messages import IMessageStore
+from mailman.interfaces.pending import IPendings
 from mailman.utilities.datetime import now
 from public import public
 from sqlalchemy import Boolean, Column, DateTime, Integer
@@ -77,6 +80,20 @@ class BounceProcessor:
     @dbconnection
     def register(self, store, mlist, email, msg, where=None):
         """See `IBounceProcessor`."""
+        # Save the DSN in the message store for notices.  It doesn't
+        # matter if we save it more than once.  Only one copy will be
+        # saved, but ensure it has a Message-ID so we can retreive it.
+        if msg.get('message-id') is None:
+            msg['Message-ID'] = make_msgid          # pragma: nocover
+        getUtility(IMessageStore).add(msg)
+        # We also need to pend a token for this or the message will be
+        # removed as an orphan by the task runner.  We don't need much
+        # from this.  We pend the msgid as _mod_message_id for the
+        # task runner.
+        pendable = _ProbePendable(
+            _mod_message_id=msg.get('message-id'))
+        getUtility(IPendings).add(
+            pendable, lifetime=as_timedelta(PENDABLE_LIFETIME))
         event = BounceEvent(mlist.list_id, email, msg, where)
         store.add(event)
         return event
@@ -200,8 +217,10 @@ class BounceProcessor:
         if member.bounce_score >= mlist.bounce_score_threshold:
             # Save bounce_score because sending probe resets it.
             saved_bounce_score = member.bounce_score
+            # Try to get the dsn from the message store.  It should be there.
+            msg = getUtility(IMessageStore).get_message_by_id(event.message_id)
             if as_boolean(config.mta.verp_probes):
-                send_probe(member, message_id=event.message_id)
+                send_probe(member, msg=msg, message_id=event.message_id)
                 action = 'sending probe'
             else:
                 self._disable_delivery(mlist, member, event)